+++ app/src/main/java/com/kwasow/musekit/MainApplication.kt
@@ -5,6 +5,7 @@ import com.kwasow.musekit.utils.MusekitPreferences
 import com.kwasow.musekit.utils.ThemeUtils
 
 class MainApplication : Application() {
+    // ====== Interface methods
     override fun onCreate() {
         super.onCreate()
