+++ app/src/main/java/com/kwasow/musekit/adapters/MainPagerAdapter.kt
@@ -10,8 +10,10 @@ import com.kwasow.musekit.fragments.SettingsFragment
 
 class MainPagerAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
+    // ====== Fields
     private val pageCount = 3
 
+    // ====== Interface methods
     override fun getItemCount(): Int = pageCount
 
     override fun createFragment(position: Int): Fragment {
