+++ app/src/main/java/com/kwasow/musekit/adapters/SoundsAdapter.kt
@@ -11,6 +11,7 @@ import com.kwasow.musekit.R
 class SoundsAdapter(context: Context, names: List<String>) :
     ArrayAdapter<String>(context, 0, names) {
 
+    // ====== Interface methods
     override fun getView(position: Int, convertView: View?, parent: ViewGroup): View {
         val soundName = getItem(position)
         val root =
