+++ app/src/main/java/com/kwasow/musekit/data/Preset.kt
@@ -1,0 +1,8 @@
+package com.kwasow.musekit.data
+
+data class Preset(
+    val name: String,
+    val semitones: Int,
+    val octave: Int,
+    val pitch: Int
+)
