+++ app/src/main/java/com/kwasow/musekit/utils/ThemeUtils.kt
@@ -1,21 +1,18 @@
 package com.kwasow.musekit.utils
 
 import android.app.Application
-import android.content.Context
-import android.content.SharedPreferences
 import android.os.Build
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.core.content.edit
 import com.google.android.material.color.DynamicColors
-import com.kwasow.musekit.R
 
 object ThemeUtils {
-    // ====== Public methods
+    // ====== Constructors
     fun init(application: Application) {
         DynamicColors.applyToActivitiesIfAvailable(application)
         AppCompatDelegate.setDefaultNightMode(getNightMode())
     }
 
+    // ====== Public methods
     fun getDefaultNightMode(): Int {
         return if (Build.VERSION.SDK_INT >= 29) {
             AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM
